requirements.txt
flask==1.1.2
pandas
joblib
pickle5
tmdbv3api==1.7.1
scikit-surprise==1.1.1
sklearn
SQLAlchemy
psycopg2-binary
jinja2<3.1.0
itsdangerous==2.0.1
Werkzeug==1.0.1
pyyaml